Assignment mode (seems to) work, forward by sample and backwards by bank
Goes between unassigned and assigned samples without crashing.
Both forward and backwards assignment motions use the same wrapper, which is called by setting a flag, and the main loop checks for the flag.
Realized that all access to the SD card must be at the same interrupt level, or else we will get an error when, for example, trying to scan the card for unassigned files at the same time a file is being played back. For now, all functions that access the sd card must be in the main loop.
Also tried setting _USE_LFN in ffconf.h to 2, which uses the stack for ff's buffers... might help some weird bugs we've been having?

Got rid of the index_bank_path[]

Also made some string functions case-insensitive (should be a separate commit, but oops)