Pitch slew when used with sequencer mostly fixed.
PITCH adc values are processes separately in a regularly run routine.
Switched PITCH to an arithmetic LPF, which reduces slew
A structure is set up for delaying the cvadc_buffer values for pitch, but it's currently set to size=1 (no delay)
When receiving a play trigger, it fades down the audio, and latches the cv adc value for a period, and inserts a delay until play is restarted. This delay is set to be enough for the LPF to stablize
ToDo: If LOOP is on when using a CV/Gate sequencer, we get bad artifacts. We probably need to disable re-triggering when the playtrig delay is in effect.
ToDo: Repeat this for other CV, especially start_pos