Fix clicks when triggering over a sample that's playing
The problem arises when the sample is still loud and a new trigger is received.
When the new trigger is received, it immediately goes to FADEDOWN mode, while waiting for the trigger delay period. However, now that the fade-down time can be as long as 50ms, the sample might still be fading down
when the trigger delay period ends.

The fix is that after trigger delay, do a RETRIG_FADEDOWN to bring audio to 0 before playing.
This adds 0.333ms of latency when triggering over already playing audio