Fix codec not recording at higher sample rates (wrong playback pitch)
Was an initialization order issue. We have to read global_params before starting codec, because we need the record_sample_rate parameter in order to know what sample rate to run the codec.

This fixes the issue where playing a sample recorded on the computer would play at the wrong pitch.
The codec was always booting at 44.1kHz, but it was doing resampling based on the assumption the codec was running at `global_params.record_sample_rate`. So if you switched to 96kHz, the codec would too, and all would be OK. But when you restarted, the codec would be at 44.1k but the module would resample as if it was at 96kHz.